src/main/java/com/jwtdemo/domain/user/UserController.java
package com.jwtdemo.domain.user;

import lombok.RequiredArgsConstructor;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RequiredArgsConstructor
@RestController
@RequestMapping("/api/v1/user")
public class UserController {

    @GetMapping("/status")
    public String status(@AuthenticationPrincipal User user) {
        return "User: " + user;
    }
}
